requirements.txt
flask
flask-restx
werkzeug
mysql-connector-python

# test
pytest

